Remove fdvar_forceEqInline
=============================== Coverage summary ================
Statements   : 97.4% ( 3074/3156 ), 75 ignored
Branches     : 93.49% ( 1437/1537 ), 76 ignored
Functions    : 98.98% ( 292/295 ), 6 ignored
Lines        : 97.55% ( 2510/2573 )
=================================================================
